Added 'constructing_profiles' and 'get_user_profile' functions.
The 'constructing_profiles' function uses the 'profiles.csv' file to make a proper profile for every user in the social network. An adjacency list is returned which consists of all the profiles. The 'get_user_profile' function returns the profile of a user whose name is given as the input.