Added 'get_recommended_friends' function along with 3 helper functions.
The 'get_profile_in_list' function takes a user's name as an input and returns their profile in a list form. The 'is_empty' function checks if the input list is empty or not. The 'get_neighbours' function takes a node as an input and returns its adjacent nodes. The 'get_recommended_friends' function takes a user's name as an input and returns their recommended friends in order from most recommended to least recommended on the basis of their mutual friends, age, country, department and hobbies.