Added 'add_friend' function and 'unfriend' function along with a helper function.
The 'get_cost' function takes a user's name as an input and returns a dictionary which consists of the cost of reaching every other person in the social network from the input user. The 'add_friend' function takes a person's name as an input (friend) and adds them to the user's friend list in the social network. The 'unfriend' function takes a person's name as an input (remove_friend) and removes them from the user's friend list in the social network.